Move to using our latest build infrastructure
This moves to using managed versions and a library catalog, this means that we get the
managed flyway version in our catalog toml file exported by the bom.

This is the only change, no other dependencies have changed, so I believe this can
go out as a patch release.

Fixes #304