Improve the Flyway configuration experience (#533)
FlywayConfigurationProperties is modified to not generate
ConfigurationBuilder properties for those FluentConfiguration properties
with ambiguous overloaded methods. Those properties are instead directly
added to FlywayConfigurationProperties and passed through to the correct
FlyentConfiguration builder method.

A FlywayConfigurationCustomizer interface is added to allow easy
programmatic additions to the configuration, allowing the user to set
properties that FlywayConfigurationProperties might not yet support or
that involve more complex Flyway types.

A default implementation of FlywayConfigurationCustomizer is provided
that looks for beans of known Flyway types with an appropriate Named
qualifier and sets them on the FluentConfiguration builder as
appropriate.